solanaWallet/transactions_demon/requirements.txt
aio-pika==6.8.0
aiofiles==0.7.0
aiohttp==3.8.1
aiormq==3.3.1
aiosignal==1.2.0
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
charset-normalizer==2.0.8
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pamqp==2.3.0
python-dotenv==0.19.2
typing-extensions==4.0.0
yarl==1.7.2
